refactor: consolidate IPC types to shared/types
- Deleted duplicate electron/ipc/types.ts (309 lines)
- Added missing payload types to shared/types/ipc.ts
- Updated 8 files to import from consolidated location
- Added PR detection payloads (PRDetectedPayload, PRClearedPayload)
- Fixed AppState type (sidebarWidth required, added lastDirectory)
- Fixed TerminalSpawnOptions (cols/rows required, cwd optional, removed unsupported args field)
- Updated shared/types/index.ts to re-export all new types
